pyproject.toml
[tool.poetry]
name = "gringo"
version = "0.1.0"
description = ""
authors = ["TheHormat <thehormat@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
python-dotenv = "^1.0.1"
alembic = "^1.14.0"
psycopg2 = "^2.9.10"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = "^3.3.0"
bcrypt = "4.0.1"
google-auth = "^2.37.0"
httpx = "^0.28.1"
google-auth-oauthlib = "^1.2.1"
google-auth-httplib2 = "^0.2.0"
requests = "^2.32.3"
requests-oauthlib = "^2.0.0"
pydantic = {extras = ["email"], version = "^2.10.5"}


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
